More granular locking -- don't hold while broadcasting.
Two other approaches I considered, and may change to moving forward:

1. Return post-lock work from the handler function.
2. Create a mutex abstraction safe to unlock multiple times.

I've long disliked multiple exit points in a function, but go makes it
pretty OK most of the time.  It gets less OK when you want to both
guarantee you release every lock you acquire *and* release your lock
as soon as possible.  #2 would allow us to unlock as soon as we're
safe, but also unlock again (noop) on return in the normal case so in
the abnormal case (e.g. panic) we can be sure to get one.