Pushed responsibility for resch response to dispatch functions.
By moving the responsibility for responding on vbreq.resch from the
vbucket.service() loop to the dispatchFun's, the dispatchFun's can now
asynchronously complete slow operations without having to block the
vbucket.service() loop.

dispatchFun() function signature now is passed the vbreq so it has
access to vbreq.resch and there's a new helper function v.respond(..)
that dispatchFun's can use when they're ready to respond with a
gomemcached.MCResponse.

There may be a cleaner, smaller approach to this; but it's not yet
apparent.