Created a bucket interface.
This was a wonderful victory for test coverage reports, as all of the
tests were passing even though this introduced a bug that caused a
test that was expecting a panic to panic *before* the expected panic.
This showed up in the test coverage report and revealed an actual
bug (which was fixed before anyone saw it).

Viva coverage.