Save item in CAS-indexed changes tree, with pointer in key-index.
Instead of saving the item in the key-indexed tree,
we now save the item (mutation) in the CAS-indexed tree.

The key-indexed tree then points to the CAS-indexed tree,
since we now save a CAS as the value in the key-indexed tree.

This change sets us up to support a "changes-feed" mutation stream
more properly.