Collection naming to have key coll come before changes coll.
We're using carefully chosen collection names, and the fact that
gkvlite flushes collections in sorted naming order, to address a
potential flushing race.

If we flush the changes collection first, but then a concurrent
goroutine does a bucketstore.set(), updating both changes and keys
collections, and then we resume flushing the keys collection...

Then, the keys colleciton would be ahead of the changes collection
(inconsistent).

This solution idea always flush keys first, which means
that keys index on disk would likely be behind the changes stream.

A separate, later fix is needed on startup/load() to fixup the keys
index.  In this approach, the changes collection is considered the
"source of truth", and the keys collection is just another secondary
index that might be outdated.