Buffer the mutation log channel.
This actually caused a deadlock-like-thing when I shifted to the new
muxer, as the subscribers were blocked waiting for a message to be
delivered that was stuck in the mutation logger.

Why something was waiting on the mutation logger was a bit of a
mystery, but it still seemed like something that should be buffered.
Previously, all the muxes were buffered at the bucket level, so it's
quite likely that that was where these messages ended up being.