Initial implementation of automatic bucket passivation.
Primarily, you get not-my-vbucket errors when the client stops
talking.  The connection holds a defunct bucket reference directly, so
when the bucket closes, the server doesn't know, and it doesn't know
which bucket it had open so it can't reconnect.

This can be solved, but it's sort of future commit status.

cbugg: close bug-472