Removed range splitting.
This makes the code faster, smaller and more agile without affecting
any of our goals.  We need to do more of this.

Note that this caused a decrease in test coverage.  This implies that
our coverage numbers were inflated by having better coverage on code we
didn't need vs. code we do.  This is also important information.

Coverage stats:

  old:  74.05% (2217/2994)
  new:  72.99% (2113/2895)