Fix field ordering so atomics align properly.
The worst part of this code is that it's almost all vbucket junk that
doesn't need to exist.

The second worst part is that it's got fragile tests that depend on the
exact ordering of the struct fields in literal JSON in the test table to
verify results.  I could've probably fixed this better, but I'd still
rather just remove all the vbucket code because it doesn't belong here
and I'm getting really tired of fixing bugs in code we don't need.

cbugg: close bug-557