Handle view query startkey=[] as nil.
A startkey of [] is json.Unmarshal()'ed into a []interface{}(nil).
But, []interface{}(nil) is json.Marshal()'ed into the string "null",
which will compare incorrectly.

This special case is to treat "null" JSON string as a golang nil so
that the view query startkey comparison works as expected.

cbugg: close bug-566