Honor stale=update_after
If we were aiming for more of a "normal" scale, I'd have each view have
a goroutine with a channel for synchronizing around view updates.  A
regular view update would send in a response channel it could wait on
that would let it know the view had been processed.  update_after would
send in a buffered response channel and just drop it.

The application would be similar, though -- a deferred message send
letting the view engine pick it up after the request completed.

But because of how we have these things come and go, I did a plain ol'
ephemeral goroutine spawned right at the end of the request for the
view.

cbugg: close bug-591