Adjust sensitivity of periodic timing test
This should never fail anywhere with a working clock except under an
extreme load.  With a timer at 1ms and a timeout of 50ms, you'd think it
wouldn't ever have failed, but we have a report of it failing after
about 90ms.

Turning the timeout up to 1s shouldn't affect the duration of the test,
but should at least make it work better on more of a starved system.

This is the only test that requires actual wall clock times because it
tests that it works with a wall clock.